+++ lightly_studio_view/src/lib/hooks/useInstanceSegmentationPreview/previewRenderer.ts
